Before optimizing

+++ day12/adventofcode2019-day12/Day12Part1/Moon.cs
@@ -18,6 +18,6 @@ namespace adventofcode2019_day12.Day12Part1
         public Position Position { get; private set; }
         public Velocity Velocity { get; private  set; }
 
-        public int Energy => (Position.Energy * Velocity.Energy);
+        public long Energy => (Position.Energy * Velocity.Energy);
     }
 }
