+++ day12/adventofcode2019-day12/Day12Part1/Position.cs
@@ -8,6 +8,6 @@ namespace adventofcode2019_day12.Day12Part1
         public int Y { get; set; }
         public int Z { get; set; }
 
-        public int Energy => (Math.Abs(X) + Math.Abs(Y) + Math.Abs(Z));
+        public long Energy => (Math.Abs(X) + Math.Abs(Y) + Math.Abs(Z));
     }
 }
